+++ scm-uaa-common/src/main/java/ir/daneshrefah/scm/uaa/common/utils/AuthenticationUtils.java
@@ -202,7 +202,7 @@ public class AuthenticationUtils {
         IpAddressMatcher ipAddressMatcher = new IpAddressMatcher(ipAddress);
         return ipAddressMatcher.matches(matchesIpAddress);
     }
-    public static Integer getLoggedInUserId() {
+    public static Long getLoggedInUserId() {
         User loggedInUser = AuthenticationUtils.getLoggedInUser();
         assert loggedInUser != null;
         return loggedInUser.getPerson().getId();
